+++ lib/bright_serializer.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require 'active_support/core_ext/string/inflections'
+require 'active_support/core_ext/array/extract_options'
 require 'bright_serializer/version'
 
 module BrightSerializer